+++ src/org/jcodings/util/CaseInsensitiveBytesHash.java
@@ -60,7 +60,7 @@ public final class CaseInsensitiveBytesHash<V> extends Hash<V>{
 
     public static int hashCode(byte[]bytes, int p, int end) {
         int key = 0;
-        while (p < end) key = ((key << 16) + (key << 6) - key) + (int)(ToLowerCaseTable[bytes[p++] & 0xff]); // & 0xff ? we have to match jruby string hash
+        while (p < end) key = ((key << 16) + (key << 6) - key) + ToLowerCaseTable[bytes[p++] & 0xff]; // & 0xff ? we have to match jruby string hash
         key = key + (key >> 5);
         return key;
     }
